+++ libevmasm/EVMAssemblyStack.h
@@ -26,6 +26,7 @@
 
 #include <map>
 #include <string>
+#include <utility>
 
 namespace solidity::evmasm
 {
@@ -33,8 +34,15 @@ namespace solidity::evmasm
 class EVMAssemblyStack: public AbstractAssemblyStack
 {
 public:
-	explicit EVMAssemblyStack(langutil::EVMVersion _evmVersion, std::optional<uint8_t> _eofVersion):
-		m_evmVersion(_evmVersion), m_eofVersion(_eofVersion) {}
+	explicit EVMAssemblyStack(
+		langutil::EVMVersion _evmVersion,
+		std::optional<uint8_t> _eofVersion,
+		Assembly::OptimiserSettings _optimiserSettings
+	):
+		m_evmVersion(_evmVersion),
+		m_eofVersion(_eofVersion),
+		m_optimiserSettings(std::move(_optimiserSettings))
+	{}
 
 	/// Runs parsing and analysis steps.
 	/// Multiple calls overwrite the previous state.
@@ -82,6 +90,7 @@ public:
 private:
 	langutil::EVMVersion m_evmVersion;
 	std::optional<uint8_t> m_eofVersion;
+	Assembly::OptimiserSettings m_optimiserSettings;
 	std::string m_name;
 	std::shared_ptr<evmasm::Assembly> m_evmAssembly;
 	std::shared_ptr<evmasm::Assembly> m_evmRuntimeAssembly;