+++ Changelog.md
@@ -6,6 +6,7 @@ Language Features:
 Compiler Features:
  * Error Reporting: Errors reported during code generation now point at the location of the contract when more fine-grained location is not available.
  * EVM: Support for the EVM version "Osaka".
+ * EVM Assembly Import: Allow enabling opcode-based optimizer.
  * SMTChecker: Support `block.blobbasefee` and `blobhash`.
  * SMTChecker: Z3 is now a runtime dependency, not a build dependency (except for emscripten build).
  * Yul Parser: Make name clash with a builtin a non-fatal error.